IGNITE-27535 Move BinarySerializedFieldComparator to test
