+++ modules/unsafe/src/main/java/org/apache/ignite/internal/pagemem/PageUtils.java
@@ -223,4 +223,37 @@ public class PageUtils {
 
         GridUnsafe.putLong(addr + off, v);
     }
+
+    /**
+     * @param addr Write address.
+     * @param type Object type.
+     * @param valBytes Value bytes array.
+     * @return Offset shift compared to initial address.
+     */
+    public static int putValue(long addr, byte type, byte[] valBytes) {
+        return putValue(addr, type, valBytes, 0, valBytes.length);
+    }
+
+    /**
+     * @param addr Write address.
+     * @param type Object type.
+     * @param srcBytes Source value bytes array.
+     * @param srcOff Start position in sourceBytes.
+     * @param len Number of bytes for write.
+     * @return Offset shift compared to initial address.
+     */
+    public static int putValue(long addr, byte type, byte[] srcBytes, int srcOff, int len) {
+        int off = 0;
+
+        PageUtils.putInt(addr, off, len);
+        off += 4;
+
+        PageUtils.putByte(addr, off, type);
+        off++;
+
+        PageUtils.putBytes(addr, off, srcBytes, srcOff, len);
+        off += len;
+
+        return off;
+    }
 }
